Fix critical/high/medium findings from fresh exhaustive review
Addresses the 2026-06-17 multi-agent codebase review (review_2026-06-17.md).

CRITICAL (5): ROCm adaptive_maxpool2d plane-local index (OOB write fix);
Vulkan deformable_conv2d bias guard; Vulkan roi_align ROI batch-idx bounds
(all dtype variants + backward); MPS Int64 index conversion; MPS softmax/
logsoftmax negative-dim normalization.

HIGH: CUDA where/masked_fill non-bool, LayerNorm-bwd F16/BF16 widen, RMSNorm
float4 alignment gate, BatchNorm F16 mean/var, cuFFT handle leak
(cufftMakePlan*), argsort non-pow2 (CUB). ROCm det 0-D scalar, BatchNorm F16
backward widen, SVD/QR/eigh/eig dtype-narrow, RMSNorm F64 native, take_along_dim
validation. Vulkan reduction-class contiguity, sort digit-order, bf16 reduction
workgroups, quantized-conv zero-point, conv-transpose bf16, layernorm bf16
backward, index negative-normalization. Core: addcmul/addcdiv/lerp broadcast,
LayerNorm multi-dim grad reshape, T5 attention scaling, RoBERTa position ids,
nested-dropout attr key, DistributedSampler drop_last, ColorJitter aliasing,
NCCL/MPI reduce_scatter AVG, JIT control-flow dtype, dequantize per-channel,
FP8-FNUZ promote_types, ModelHub retry truncation.

MEDIUM (CPU): where_batch NaN predicate, logsumexp isinf on accumulator,
SliceScatter off-by-one, topk k<0 guard, activation-backward contiguity,
one_hot OOB throw, BatchNorm F16 forward float-accum.

OneAPI and MPS fixes are source-only (no toolchain on this host) and unverified;
all other fixes built and passed targeted op-level tests with zero failures.

+++ src/backends/cpu/kernels/reduction.cpp
@@ -4088,7 +4088,7 @@ static void logsumexp_along_dim(
         }
 
         // Handle inf
-        if (std::isinf(static_cast<float>(m))) {
+        if (std::isinf(m)) {  // test on the accumulator type; casting double->float spuriously overflows finite large Float64
             output_data[out_idx] = T(m);
             continue;
         }
